Add files via upload
I found this a challenging exercise  given how new I am to coding. It involved another language that I a not familiar with "Regular Expression" and the aim was to find numbers scattered though a long line to text, extract and sum them up together. I found it a good exercise on working with data structures and reading external file.